# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRT Portal API",
    "description": "API documentation for CRT Portal",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/students/{id}": {
      "get": {
        "tags": [
          "student-controller"
        ],
        "operationId": "getStudentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "student-controller"
        ],
        "operationId": "updateStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "student-controller"
        ],
        "operationId": "deleteStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sections/{sectionId}": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "operationId": "getSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "section-controller"
        ],
        "operationId": "updateSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "section-controller"
        ],
        "operationId": "deleteSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sections": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "operationId": "getAllSections",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "section-controller"
        ],
        "operationId": "createSection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/faculty/attendance": {
      "post": {
        "tags": [
          "Faculty Dashboard"
        ],
        "summary": "Submit attendance",
        "description": "Submit attendance for a time slot with topic taught",
        "operationId": "submitAttendance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceSubmissionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSubmissionResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/faculty/students/{sectionId}": {
      "get": {
        "tags": [
          "Faculty Dashboard"
        ],
        "summary": "Get students for section",
        "description": "Returns list of students in a specific section for attendance",
        "operationId": "getStudentsForSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudentListDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/faculty/session/{timeSlotId}/students": {
      "get": {
        "tags": [
          "Faculty Dashboard"
        ],
        "summary": "Get students for time slot",
        "description": "Returns list of students for a specific time slot session",
        "operationId": "getStudentsForTimeSlot",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudentListDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StudentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "regNum": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "batch": {
            "type": "string",
            "enum": [
              "Y22",
              "Y23",
              "Y24",
              "Y25"
            ]
          },
          "crtEligibility": {
            "type": "boolean"
          },
          "feedback": {
            "type": "string"
          },
          "attendancePercentage": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CreateSectionDTO": {
        "type": "object",
        "properties": {
          "trainerId": {
            "type": "string",
            "format": "uuid"
          },
          "sectionName": {
            "type": "string"
          }
        }
      },
      "CRT_TrainerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "SectionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "trainer": {
            "$ref": "#/components/schemas/CRT_TrainerDTO"
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentDTO"
            },
            "uniqueItems": true
          },
          "strength": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AttendanceSubmissionDTO": {
        "type": "object",
        "properties": {
          "timeSlotId": {
            "type": "string"
          },
          "sectionId": {
            "type": "string"
          },
          "topicTaught": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "attendanceRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentAttendanceRecordDTO"
            }
          }
        }
      },
      "StudentAttendanceRecordDTO": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "feedback": {
            "type": "string"
          },
          "present": {
            "type": "boolean"
          }
        }
      },
      "AttendanceSessionResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "facultyId": {
            "type": "string"
          },
          "sectionId": {
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          },
          "topicTaught": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "timeSlot": {
            "type": "string"
          },
          "room": {
            "type": "string"
          },
          "totalStudents": {
            "type": "integer",
            "format": "int32"
          },
          "presentCount": {
            "type": "integer",
            "format": "int32"
          },
          "absentCount": {
            "type": "integer",
            "format": "int32"
          },
          "attendancePercentage": {
            "type": "number",
            "format": "double"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AttendanceSubmissionResponseDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "attendanceSession": {
            "$ref": "#/components/schemas/AttendanceSessionResponseDTO"
          }
        }
      },
...
      "StudentListDTO": {
        "type": "object",
        "properties": {
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentDTO"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "sectionName": {
            "type": "string"
          },
          "sectionId": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}